Refactor payment calculations and add new entities
Updated the `CalculatePayment` method to return `decimal` for better precision. Added new `DbSet` properties in `DataContext.cs` for `RoomTypes`, `EmployeeTask`, `Payroll`, `Tax`, `Feedback`, `Attendance`, `Ratings`, and `YearlyRatingsSummary`. Expanded `OnModelCreating` to configure relationships for these entities. Modified related models to reflect new relationships and data types, including navigation properties for `Feedbacks` in `Customer.cs` and `Reservation.cs`. Created new models for `Attendance`, `EmployeeTask`, and `Tax`. Added migration files to update the database schema accordingly.

+++ Controllers/ReservationController.cs
@@ -163,7 +163,8 @@ namespace SolmileAPI.Controllers
         [Route("CalculatePayment/{reservationId}")]
         public async Task<IActionResult> CalculatePayment([FromRoute] string reservationId)
         {
-            float paymentAmount = await _reservationRepo.CalculatePaymentAsync(reservationId);
+            decimal paymentAmount = await _reservationRepo.CalculatePaymentAsync(reservationId);
+
             if (paymentAmount == 0)
                 return NotFound("Reservation not found or invalid data.");
