+++ Interface/ReservationInterface.cs
@@ -4,7 +4,7 @@ namespace SolmileAPI.Interface
 {
     public interface ReservationInterface
     {
-            Task<float> CalculatePaymentAsync(string reservationId);
+            Task<decimal> CalculatePaymentAsync(string reservationId);
             Task<Reservation> CreateReservationAsync( string roomId, DateTime checkIn, DateTime checkOut);
             Task<bool> UpdateReservationAsync(string reservationId, Reservation updatedDetails);
             Task<bool> UpdateReservationAsync(string reservationId, DateTime newCheckOutDate);
